Read even class labels
